+++ packages/lumx-react/src/components/chip/Chip.stories.tsx
@@ -6,7 +6,7 @@ import { getSelectArgType } from '@lumx/core/stories/controls/selectArgType';
 import { withCombinations } from '@lumx/react/stories/decorators/withCombinations';
 import { withUndefined } from '@lumx/core/stories/controls/withUndefined';
 import { withTheming } from '@lumx/react/stories/utils/theming';
-import { DESIGN_TOKENS } from '@lumx/core/src/js/constants/_internal/design-tokens';
+import { DESIGN_TOKENS } from '@lumx/core/js/constants/_internal/design-tokens';
 import pick from 'lodash/pick';
 import { withThemedBackground } from '@lumx/react/stories/decorators/withThemedBackground';